Move static inside public
