+++ public/static/media/embeddingCategorical.png
