+++ public/static/media/embeddingContinuous.png
