+++ public/static/media/heatmap.png
