+++ public/static/media/plot6.png
