+++ public/static/media/plot7.png
